Add tests for float truncation and (de)serialization
We increased the number of default bits for the float truncation
mask to 12, in order to decrease the precision of floats stored in
attributes, such that (de)serialization of values in django is
consistent. This is still failing for at least one set of test values
in `aiida.common.test_hashing`, which were generated by the skipped
test in `aiida.backends.tests.clean_value`. Note that this test also
passes when increasing the masking bits to 15.

Then in SqlAlchemy there is still a problem where large positive floats
are converted to a long and therefore returned as such.